+++ lb.tf
@@ -18,6 +18,14 @@ resource "aws_security_group" "alb_sgrp" {
   }
 }
 
+resource "aws_vpc_security_group_egress_rule" "alb_to_ecs_egress" {
+  security_group_id            = aws_security_group.alb_sgrp.id
+  from_port                    = 80
+  to_port                      = 80
+  ip_protocol                  = "tcp"
+  referenced_security_group_id = aws_security_group.ecs_sgrp.id
+}
+
 resource "aws_lb" "nginx_alb" {
   name               = "${var.service}-alb"
   internal           = false